requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
faiss-cpu>=1.7.4
pypdf>=3.17.4
pydantic>=2.5.3
requests>=2.31.0

